requirements.txt
sounddevice~=0.5.1
torch~=2.5.1
transformers~=4.46.2
opencv-python~=4.10.0.84
ultralytics~=8.3.33
paho-mqtt~=2.1.0
numpy~=2.1.3
influxdb-client~=1.48.0